+++ CHANGELOG.md
@@ -575,7 +575,9 @@
   - Added sample component `SkeletonAnimationFixedTimestepThreaded` as a replacement component for `SkeletonAnimationFixedTimestep` which was limited to single-threaded use only. This component is using the new `DeltaTimeOverride` event and is thus compatible with both single-threaded and multi-threaded animation.
   - Added sample component `SkeletonAnimationStaticPoseOptimizer` which automatically skips animation and mesh updates while a SkeletonAnimation's `AnimationState` retains a static pose. It is compatible with threaded animation and mesh updates. The component modifies `SkeletonAnimation.UpdateMode`, so watch out for side effects if multiple components modify this same property.
   - On-demand loading modules: Added method `HasRequestFailed` to `GenericOnDemandTextureLoader`, automatically handling when requests fail to be processed. It provides a default implementation returning `false` to maintain the existing behaviour without code change. `AddressableTextureLoader` implements this method accordingly to cover Addressable request failure. Be sure to update both `com.esotericsoftware.spine.addressables` and `com.esotericsoftware.spine.on-demand-loading` UPM packages.
-  
+  - On-demand loading modules: Added `IDeferredCleanupOnDemandTextureLoader` interface for automatic cleanup by `OnDemandTextureLoaderCleanup` below. `GenericOnDemandTextureLoader` implements this interface.
+  - On-demand loading modules: Added `OnDemandTextureLoaderCleanup`, which automatically maintains loaders with unreleased texture requests independently of skeleton renderer lifecycle. This is necessary because textures are deliberately kept alive for a delay after their last use to avoid unload/reload thrashing, so cleanup must continue after the last renderer using them is destroyed. When loaders are destroyed or unloaded they now also unload all textures and restore their low-resolution placeholders.
+
 - **Deprecated**
 
 - **Restructuring (Non-Breaking)**